@@ -8,6 +8,7 @@ using Volo.Abp.Identity;
 using Volo.Abp.Modularity;
 using Volo.Abp.ObjectExtending;
 using Volo.Abp.PermissionManagement;
+using Volo.Abp.SettingManagement;
 using Volo.Abp.TenantManagement;
 
 namespace Generic.Abp.Demo
@@ -18,9 +19,9 @@ namespace Generic.Abp.Demo
         typeof(AbpFeatureManagementApplicationContractsModule),
         typeof(AbpIdentityApplicationContractsModule),
         typeof(AbpPermissionManagementApplicationContractsModule),
+        typeof(AbpSettingManagementApplicationContractsModule),
         typeof(AbpTenantManagementApplicationContractsModule),
         typeof(AbpObjectExtendingModule),
-        typeof(GenericAbpEnumerationDomainSharedModule),
         typeof(GenericAbpExtResourceApplicationContractsModule),
         typeof(GenericAbpIdentityApplicationContractsModule),
         typeof(GenericAbpEnumerationApplicationContractsModule),
