@@ -1,7 +1,6 @@
 ﻿using Generic.Abp.Enumeration;
 using Generic.Abp.ExtResource;
 using Generic.Abp.Identity;
-using Generic.Abp.IdentityServer;
 using Generic.Abp.OpenIddict;
 using Volo.Abp.Account;
 using Volo.Abp.AutoMapper;
@@ -10,6 +9,7 @@ using Volo.Abp.Identity;
 using Volo.Abp.Modularity;
 using Volo.Abp.PermissionManagement;
 using Volo.Abp.TenantManagement;
+using Volo.Abp.SettingManagement;
 
 namespace Generic.Abp.Demo
 {
@@ -21,10 +21,10 @@ namespace Generic.Abp.Demo
         typeof(AbpPermissionManagementApplicationModule),
         typeof(AbpTenantManagementApplicationModule),
         typeof(AbpFeatureManagementApplicationModule),
+        typeof(AbpSettingManagementApplicationModule),
         typeof(GenericAbpEnumerationApplicationModule),
         typeof(GenericAbpExtResourceApplicationModule),
         typeof(GenericAbpIdentityApplicationModule),
-        typeof(GenericAbpEnumerationApplicationModule),
         typeof(GenericAbpOpenIddictApplicationModule)
         )]
     public class DemoApplicationModule : AbpModule
