@@ -1,28 +1,33 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Generic.Abp.Demo.Data;
 using Serilog;
 using Volo.Abp;
+using Volo.Abp.Data;
 
 namespace Generic.Abp.Demo.DbMigrator
 {
     public class DbMigratorHostedService : IHostedService
     {
         private readonly IHostApplicationLifetime _hostApplicationLifetime;
-
-        public DbMigratorHostedService(IHostApplicationLifetime hostApplicationLifetime)
+        private readonly IConfiguration _configuration;
+        public DbMigratorHostedService(IHostApplicationLifetime hostApplicationLifetime, IConfiguration configuration)
         {
             _hostApplicationLifetime = hostApplicationLifetime;
+            _configuration = configuration;
         }
 
         public async Task StartAsync(CancellationToken cancellationToken)
         {
             using (var application = AbpApplicationFactory.Create<DemoDbMigratorModule>(options =>
             {
+                options.Services.ReplaceConfiguration(_configuration);
                 options.UseAutofac();
                 options.Services.AddLogging(c => c.AddSerilog());
+                options.AddDataMigrationEnvironment();
             }))
             {
                 application.Initialize();
@@ -32,7 +37,7 @@ namespace Generic.Abp.Demo.DbMigrator
                     .GetRequiredService<DemoDbMigrationService>()
                     .MigrateAsync();
 
-                application.Shutdown();
+                await application.ShutdownAsync();
 
                 _hostApplicationLifetime.StopApplication();
             }
