@@ -7,14 +7,10 @@ namespace Generic.Abp.Demo.DbMigrator
 {
     [DependsOn(
         typeof(AbpAutofacModule),
-        typeof(DemoEntityFrameworkCoreDbMigrationsModule),
+        typeof(DemoEntityFrameworkCoreModule),
         typeof(DemoApplicationContractsModule)
         )]
     public class DemoDbMigratorModule : AbpModule
     {
-        public override void ConfigureServices(ServiceConfigurationContext context)
-        {
-            Configure<AbpBackgroundJobOptions>(options => options.IsJobExecutionEnabled = false);
-        }
     }
 }
