@@ -1,10 +1,15 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.IO;
 using System.Linq;
+using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using Volo.Abp.Data;
 using Volo.Abp.DependencyInjection;
+using Volo.Abp.Identity;
 using Volo.Abp.MultiTenancy;
 using Volo.Abp.TenantManagement;
 
@@ -35,6 +40,13 @@ namespace Generic.Abp.Demo.Data
 
         public async Task MigrateAsync()
         {
+            var initialMigrationAdded = AddInitialMigrationIfNotExist();
+
+            if (initialMigrationAdded)
+            {
+                return;
+            }
+
             Logger.LogInformation("Started database migrations...");
 
             await MigrateDatabaseSchemaAsync();
@@ -47,22 +59,20 @@ namespace Generic.Abp.Demo.Data
             var migratedDatabaseSchemas = new HashSet<string>();
             foreach (var tenant in tenants)
             {
-                if (!tenant.ConnectionStrings.Any())
-                {
-                    continue;
-                }
-
                 using (_currentTenant.Change(tenant.Id))
                 {
-                    var tenantConnectionStrings = tenant.ConnectionStrings
-                        .Select(x => x.Value)
-                        .ToList();
-
-                    if (!migratedDatabaseSchemas.IsSupersetOf(tenantConnectionStrings))
+                    if (tenant.ConnectionStrings.Any())
                     {
-                        await MigrateDatabaseSchemaAsync(tenant);
+                        var tenantConnectionStrings = tenant.ConnectionStrings
+                            .Select(x => x.Value)
+                            .ToList();
 
-                        migratedDatabaseSchemas.AddIfNotContains(tenantConnectionStrings);
+                        if (!migratedDatabaseSchemas.IsSupersetOf(tenantConnectionStrings))
+                        {
+                            await MigrateDatabaseSchemaAsync(tenant);
+
+                            migratedDatabaseSchemas.AddIfNotContains(tenantConnectionStrings);
+                        }
                     }
 
                     await SeedDataAsync(tenant);
@@ -71,7 +81,8 @@ namespace Generic.Abp.Demo.Data
                 Logger.LogInformation($"Successfully completed {tenant.Name} tenant database migrations.");
             }
 
-            Logger.LogInformation("Successfully completed database migrations.");
+            Logger.LogInformation("Successfully completed all database migrations.");
+            Logger.LogInformation("You can safely end this process...");
         }
 
         private async Task MigrateDatabaseSchemaAsync(Tenant tenant = null)
@@ -89,7 +100,124 @@ namespace Generic.Abp.Demo.Data
         {
             Logger.LogInformation($"Executing {(tenant == null ? "host" : tenant.Name + " tenant")} database seed...");
 
-            await _dataSeeder.SeedAsync(tenant?.Id);
+            await _dataSeeder.SeedAsync(new DataSeedContext(tenant?.Id)
+                .WithProperty(IdentityDataSeedContributor.AdminEmailPropertyName, IdentityDataSeedContributor.AdminEmailDefaultValue)
+                .WithProperty(IdentityDataSeedContributor.AdminPasswordPropertyName, IdentityDataSeedContributor.AdminPasswordDefaultValue)
+            );
         }
+
+        private bool AddInitialMigrationIfNotExist()
+        {
+            try
+            {
+                if (!DbMigrationsProjectExists())
+                {
+                    return false;
+                }
+            }
+            catch (Exception)
+            {
+                return false;
+            }
+
+            try
+            {
+                if (!MigrationsFolderExists())
+                {
+                    AddInitialMigration();
+                    return true;
+                }
+                else
+                {
+                    return false;
+                }
+            }
+            catch (Exception e)
+            {
+                Logger.LogWarning("Couldn't determinate if any migrations exist : " + e.Message);
+                return false;
+            }
+        }
+
+        private bool DbMigrationsProjectExists()
+        {
+            var dbMigrationsProjectFolder = GetEntityFrameworkCoreProjectFolderPath();
+
+            return dbMigrationsProjectFolder != null;
+        }
+
+        private bool MigrationsFolderExists()
+        {
+            var dbMigrationsProjectFolder = GetEntityFrameworkCoreProjectFolderPath();
+
+            return Directory.Exists(Path.Combine(dbMigrationsProjectFolder, "Migrations"));
+        }
+
+        private void AddInitialMigration()
+        {
+            Logger.LogInformation("Creating initial migration...");
+
+            string argumentPrefix;
+            string fileName;
+
+            if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX) || RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
+            {
+                argumentPrefix = "-c";
+                fileName = "/bin/bash";
+            }
+            else
+            {
+                argumentPrefix = "/C";
+                fileName = "cmd.exe";
+            }
+
+            var procStartInfo = new ProcessStartInfo(fileName,
+                $"{argumentPrefix} \"abp create-migration-and-run-migrator \"{GetEntityFrameworkCoreProjectFolderPath()}\"\""
+            );
+
+            try
+            {
+                Process.Start(procStartInfo);
+            }
+            catch (Exception)
+            {
+                throw new Exception("Couldn't run ABP CLI...");
+            }
+        }
+
+        private string GetEntityFrameworkCoreProjectFolderPath()
+        {
+            var slnDirectoryPath = GetSolutionDirectoryPath();
+
+            if (slnDirectoryPath == null)
+            {
+                throw new Exception("Solution folder not found!");
+            }
+
+            var srcDirectoryPath = Path.Combine(slnDirectoryPath, "src");
+
+            return Directory.GetDirectories(srcDirectoryPath)
+                .FirstOrDefault(d => d.EndsWith(".EntityFrameworkCore"));
+        }
+
+        private string GetSolutionDirectoryPath()
+        {
+            var currentDirectory = new DirectoryInfo(Directory.GetCurrentDirectory());
+
+            while (Directory.GetParent(currentDirectory.FullName) != null)
+            {
+                currentDirectory = Directory.GetParent(currentDirectory.FullName);
+
+                if (Directory.GetFiles(currentDirectory.FullName).FirstOrDefault(f => f.EndsWith(".sln")) != null)
+                {
+                    return currentDirectory.FullName;
+                }
+            }
+
+            return null;
+        }
+
     }
+
+
 }
