@@ -1,11 +1,17 @@
 ﻿using Microsoft.EntityFrameworkCore;
-using Generic.Abp.Demo.Users;
+using Volo.Abp.AuditLogging.EntityFrameworkCore;
+using Volo.Abp.BackgroundJobs.EntityFrameworkCore;
 using Volo.Abp.Data;
+using Volo.Abp.DependencyInjection;
 using Volo.Abp.EntityFrameworkCore;
-using Volo.Abp.EntityFrameworkCore.Modeling;
+using Volo.Abp.FeatureManagement.EntityFrameworkCore;
 using Volo.Abp.Identity;
-using Volo.Abp.Users.EntityFrameworkCore;
+using Volo.Abp.Identity.EntityFrameworkCore;
 using Volo.Abp.OpenIddict.EntityFrameworkCore;
+using Volo.Abp.PermissionManagement.EntityFrameworkCore;
+using Volo.Abp.SettingManagement.EntityFrameworkCore;
+using Volo.Abp.TenantManagement;
+using Volo.Abp.TenantManagement.EntityFrameworkCore;
 
 namespace Generic.Abp.Demo.EntityFrameworkCore
 {
@@ -18,14 +24,43 @@ namespace Generic.Abp.Demo.EntityFrameworkCore
      * Don't use this DbContext for database migrations since it does not contain tables of the
      * used modules (as explained above). See DemoMigrationsDbContext for migrations.
      */
+    [ReplaceDbContext(typeof(IIdentityDbContext))]
+    [ReplaceDbContext(typeof(ITenantManagementDbContext))]
     [ConnectionStringName("Default")]
-    public class DemoDbContext : AbpDbContext<DemoDbContext>
+    public class DemoDbContext :
+    AbpDbContext<DemoDbContext>,
+    IIdentityDbContext,
+    ITenantManagementDbContext
     {
-        public DbSet<AppUser> Users { get; set; }
 
-        /* Add DbSet properties for your Aggregate Roots / Entities here.
-         * Also map them inside DemoDbContextModelCreatingExtensions.ConfigureDemo
+        /* Add DbSet properties for your Aggregate Roots / Entities here.*/
+
+        #region Entities from the modules
+
+        /* Notice: We only implemented IIdentityDbContext and ITenantManagementDbContext
+         * and replaced them for this DbContext. This allows you to perform JOIN
+         * queries for the entities of these modules over the repositories easily. You
+         * typically don't need that for other modules. But, if you need, you can
+         * implement the DbContext interface of the needed module and use ReplaceDbContext
+         * attribute just like IIdentityDbContext and ITenantManagementDbContext.
+         *
+         * More info: Replacing a DbContext of a module ensures that the related module
+         * uses this DbContext on runtime. Otherwise, it will use its own DbContext class.
          */
+
+        //Identity
+        public DbSet<IdentityUser> Users { get; set; }
+        public DbSet<IdentityRole> Roles { get; set; }
+        public DbSet<IdentityClaimType> ClaimTypes { get; set; }
+        public DbSet<OrganizationUnit> OrganizationUnits { get; set; }
+        public DbSet<IdentitySecurityLog> SecurityLogs { get; set; }
+        public DbSet<IdentityLinkUser> LinkUsers { get; set; }
+
+        // Tenant Management
+        public DbSet<Tenant> Tenants { get; set; }
+        public DbSet<TenantConnectionString> TenantConnectionStrings { get; set; }
+
+        #endregion
 
         public DemoDbContext(DbContextOptions<DemoDbContext> options)
             : base(options)
@@ -37,23 +72,26 @@ namespace Generic.Abp.Demo.EntityFrameworkCore
         {
             base.OnModelCreating(builder);
 
-            /* Configure the shared tables (with included modules) here */
+            /* Include modules to your migration db context */
 
-            builder.Entity<AppUser>(b =>
-            {
-                b.ToTable(AbpIdentityDbProperties.DbTablePrefix + "Users"); //Sharing the same table "AbpUsers" with the IdentityUser
-                
-                b.ConfigureByConvention();
-                b.ConfigureAbpUser();
+            builder.ConfigurePermissionManagement();
+            builder.ConfigureSettingManagement();
+            builder.ConfigureBackgroundJobs();
+            builder.ConfigureAuditLogging();
+            builder.ConfigureIdentity();
+            builder.ConfigureOpenIddict();
+            builder.ConfigureFeatureManagement();
+            builder.ConfigureTenantManagement();
 
-                /* Configure mappings for your additional properties
-                 * Also see the DemoEfCoreEntityExtensionMappings class
-                 */
-            });
+            /* Configure your own tables/entities inside here */
 
-            /* Configure your own tables/entities inside the ConfigureDemo method */
+            //builder.Entity<YourEntity>(b =>
+            //{
+            //    b.ToTable(BookStoreConsts.DbTablePrefix + "YourEntities", BookStoreConsts.DbSchema);
+            //    b.ConfigureByConvention(); //auto configure for the base class props
+            //    //...
+            //});
 
-            builder.ConfigureDemo();
         }
     }
 }
