@@ -5,6 +5,8 @@ using Generic.Abp.Demo.Localization;
 using Generic.Abp.Demo.MultiTenancy;
 using Volo.Abp.TenantManagement.Web.Navigation;
 using Volo.Abp.UI.Navigation;
+using Volo.Abp.Identity.Web.Navigation;
+using Volo.Abp.SettingManagement.Web.Navigation;
 
 namespace Generic.Abp.Demo.Web.Menus
 {
@@ -20,15 +22,31 @@ namespace Generic.Abp.Demo.Web.Menus
 
         private Task ConfigureMainMenuAsync(MenuConfigurationContext context)
         {
-            if (!MultiTenancyConsts.IsEnabled)
+            var administration = context.Menu.GetAdministration();
+            var l = context.GetLocalizer<DemoResource>();
+
+            context.Menu.Items.Insert(
+                0,
+                new ApplicationMenuItem(
+                    DemoMenus.Home,
+                    l["Menu:Home"],
+                    "~/",
+                    icon: "fas fa-home",
+                    order: 0
+                )
+            );
+
+            if (MultiTenancyConsts.IsEnabled)
             {
-                var administration = context.Menu.GetAdministration();
+                administration.SetSubItemOrder(TenantManagementMenuNames.GroupName, 1);
+            }
+            else
+            {
                 administration.TryRemoveMenuItem(TenantManagementMenuNames.GroupName);
             }
 
-            var l = context.GetLocalizer<DemoResource>();
-
-            context.Menu.Items.Insert(0, new ApplicationMenuItem("Demo.Home", l["Menu:Home"], "~/"));
+            administration.SetSubItemOrder(IdentityMenuNames.GroupName, 2);
+            administration.SetSubItemOrder(SettingManagementMenuNames.GroupName, 3);
 
             return Task.CompletedTask;
         }
