@@ -11,9 +11,9 @@
 @inject IAuthorizationService _authorization
 @inject IPageLayout _pageLayout
 @{
-    _pageLayout.Content.Title = _l["Scopes"].Value;
-    _pageLayout.Content.BreadCrumb.Add(_l["Menu:OpenIddictManagement"].Value);
-    _pageLayout.Content.MenuItemName = OpenIddictMenus.Scopes;
+    //_pageLayout.Content.Title = _l["Scopes"].Value;
+    //_pageLayout.Content.BreadCrumb.Add(_l["Menu:OpenIddictManagement"].Value);
+    //_pageLayout.Content.MenuItemName = OpenIddictMenus.Scopes;
 }
 @section styles{
     <abp-style-bundle name="W2Ui"></abp-style-bundle>
